Create data package, move grpc under it, utils reorg
