+++ pinecone/__init__.py
@@ -1,12 +1,12 @@
 """
 .. include:: ../README.md
 """
-from .config.config import *
+from .config import *
 from .exceptions import *
 from .control.pinecone import Pinecone
-from .index import *
+from .data.index import *
 
 try:
-    from .grpc.index_grpc import *
+    from .data.grpc.index_grpc import *
 except ImportError:
     pass  # ignore for non-[grpc] installations
