+++ pinecone/data/__init__.py
@@ -1,0 +1,2 @@
+from .index import Index
+from .grpc.index_grpc import GRPCIndex
